Turn controls into controller class
